+++ app/src/main/java/ch/epfl/sdp/ui/maps/offline/MapsManagingFragment.kt
@@ -10,7 +10,6 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import ch.epfl.sdp.MainApplication
 import ch.epfl.sdp.R
 import ch.epfl.sdp.map.offline.OfflineRegionUtils
-import ch.epfl.sdp.ui.mapsManaging.MapSelectionRecyclerViewAdapter
 import com.mapbox.mapboxsdk.offline.OfflineManager
 import com.mapbox.mapboxsdk.offline.OfflineRegion
 import kotlinx.android.synthetic.main.fragment_maps_managing.*
